rename eventmesh to eventing, update dependencies
