+++ tests/end-to-end/upgrade/Gopkg.toml
@@ -40,10 +40,6 @@ required = [
   version = "6.15.2"
 
 [[constraint]]
-  name = "k8s.io/cli-runtime"
-  branch = "release-1.12"
-
-[[constraint]]
   name = "github.com/gofrs/uuid"
   version = "3.2.0"
 
@@ -54,21 +50,6 @@ required = [
 [[constraint]]
   name = "github.com/avast/retry-go"
   version = "2.2.0"
-
-[[constraint]]
-  name = "knative.dev/eventing"
-  version = "0.12.0"
-
-# Transitive dependencies of knative.dev/pkg
-[[override]]
-  name = "k8s.io/api"
-  version = "kubernetes-1.15.3"
-[[override]]
-  name = "k8s.io/apimachinery"
-  version = "kubernetes-1.15.3"
-[[override]]
-  name = "k8s.io/client-go"
-  version = "kubernetes-1.15.3"
 
 [[override]]
   name = "golang.org/x/crypto"
