+++ tests/end-to-end/upgrade/pkg/tests/eventing/setup.go
@@ -1,0 +1,7 @@
+package eventing
+
+const (
+	defaultName             = "eventupgrade"
+	defaultEventType        = "order.created"
+	defaultEventTypeVersion = "v1"
+)
