+++ tests/end-to-end/upgrade/pkg/tests/eventing/subscription/api/v1alpha1/groupversion_info.go
