+++ tests/end-to-end/upgrade/pkg/tests/eventing/subscription/api/v1alpha1/subscription_lifecycle.go
