+++ tests/end-to-end/upgrade/pkg/tests/eventing/subscription/api/v1alpha1/zz_generated.deepcopy.go
