+++ tests/end-to-end/upgrade/pkg/tests/service-catalog/report.go
@@ -11,7 +11,6 @@ import (
 	sbu "github.com/kyma-project/kyma/components/service-binding-usage-controller/pkg/apis/servicecatalog/v1alpha1"
 	appsV1 "k8s.io/api/apps/v1"
 	coreV1 "k8s.io/api/core/v1"
-	ch "knative.dev/eventing/pkg/apis/messaging/v1alpha1"
 
 	"github.com/sirupsen/logrus"
 )
@@ -68,12 +67,6 @@ type (
 )
 
 type (
-	Channel struct {
-		Status ch.ChannelStatus
-	}
-)
-
-type (
 	Log struct {
 		Message string
 	}
@@ -105,8 +98,6 @@ type Report struct {
 	ServiceBindingUsages  map[string]ServiceBindingUsage `json:"service_binding_usages"`
 
 	Applications map[string]Application `json:"applications"`
-
-	Channels map[string]Channel `json:"channels"`
 }
 
 func NewReport(namespace string, log logrus.FieldLogger) *Report {
@@ -127,7 +118,6 @@ func NewReport(namespace string, log logrus.FieldLogger) *Report {
 		ServiceBindings:       make(map[string]ServiceBinding, 0),
 		ServiceBindingUsages:  make(map[string]ServiceBindingUsage, 0),
 		Applications:          make(map[string]Application, 0),
-		Channels:              make(map[string]Channel, 0),
 	}
 }
 
@@ -362,23 +352,6 @@ func (r *Report) AddApplications(applications *appOpr.ApplicationList, err error
 		}
 	}
 	r.Details["Application"] = fmt.Sprintf("%s (%s)", r.Details["Application"], strings.Join(names, ", "))
-}
-
-func (r *Report) AddChannels(channels *ch.ChannelList, err error) {
-	if err != nil {
-		r.Warnings["Channel"] = fmt.Sprintf("cannot fetch Channels list: %s", err)
-		return
-	}
-
-	r.Details["Channel"] = fmt.Sprintf("report found %d Channels", len(channels.Items))
-	names := make([]string, 0)
-	for _, channel := range channels.Items {
-		names = append(names, channel.Name)
-		r.Channels[channel.Name] = Channel{
-			Status: channel.Status,
-		}
-	}
-	r.Details["Channel"] = fmt.Sprintf("%s (%s)", r.Details["Channel"], strings.Join(names, ", "))
 }
 
 func (r *Report) AddLogs(name string, logs []string, grepWords []string, err error) {
